Initial separation of the two RedFish mock scripts, one for the server and one to create the CSV file for replay
