+++ jsMockRedfishServer/rfmockserver.js
@@ -1,0 +1,122 @@
+"use strict";
+
+var util = require("util");
+var fs = require("fs");
+
+var config;
+var PFuncs = require("./PatternFuncs");
+
+if (process.argv.indexOf("-c") != -1) {
+   if (process.argv[process.argv.indexOf("-c") + 1] != -1) {
+      var configFile = "./" + process.argv[process.argv.indexOf("-c") + 1];
+      console.log("Using config file: " + configFile);
+      try {
+         config = require(configFile);
+      } catch (e) {
+         console.log("Error opening " + configFile + ": " + e);
+      }
+   }
+}
+
+if (!config) {
+   config = require("./config");
+}
+
+if (process.argv.indexOf("-s") != 1 && process.argv.indexOf("-f") == -1) {
+   startServer(config.RedFishData.path);
+}
+
+function startServer(redfishPath) {
+   var express = require("express");
+   var app = express();
+
+   app.use("/redfish", express.static(redfishPath, { index: "index.json" }));
+
+   app.listen(8001);
+
+   console.log("Server started on port 8001");
+}
+
+var patternTimers = [];
+
+config.MockupData.MockupPatterns.forEach(function(mockup, index) {
+  var p = new PFuncs();
+
+  if (!typeof p[mockup.pattern]) {
+     console.log(
+        "Pattern '" +
+           mockup.pattern +
+           "' not implemented! Skipping " +
+           mockup.name +
+           "."
+     );
+     return;
+  }
+
+  if (!patternTimers[index]) patternTimers[index] = {};
+
+  if (mockup.min) {
+     p.min = mockup.min;
+  }
+  if (mockup.max) {
+     p.max = mockup.max;
+  }
+  if (mockup.step) {
+     p.step = mockup.step;
+  }
+
+  patternTimers[index].pfuncs = p;
+
+  patternTimers[index].timer = setInterval(function() {
+     var isoDTG = new Date().toISOString();
+     isoDTG = isoDTG.replace(/\..*$/, "-0500");
+
+     var currJSON = JSON.parse(
+        fs.readFileSync(
+           config.RedFishData.path + mockup.path + "index.json",
+           "utf-8"
+        )
+     );
+
+     var templateJSON = JSON.stringify(mockup.MetricValueTemplate);
+
+     templateJSON = templateJSON.replace(
+        /#value/g,
+        patternTimers[index].pfuncs[mockup.pattern]()
+     );
+     templateJSON = templateJSON.replace(/#timestamp/g, isoDTG);
+
+     var template = JSON.parse(templateJSON);
+
+     var tmpMVs = currJSON.MetricValues.filter(function(mval) {
+        return (
+           mval.MemberID !== template.MemberID ||
+           mval.MetricProperty !== template.MetricProperty
+        );
+     });
+
+     currJSON.MetricValues = tmpMVs;
+
+     currJSON.MetricValues.push(template);
+
+     console.log(
+        isoDTG +
+           ": " +
+           mockup.name +
+           "(" +
+           patternTimers[index].pfuncs.value +
+           ")"
+     );
+
+     fs.writeFileSync(
+        config.RedFishData.path + mockup.path + "index.json",
+        JSON.stringify(currJSON, null, "\t"),
+        "utf-8",
+        function(err) {
+           if (err) {
+              return console.log(err);
+           }
+        }
+     );
+  }, mockup.timedelay * 1000);
+});
